Fix WebTransport tunnels
Close wasn't working properly, which caused video streaming in
the browser to hang when seeking around to multiple places in
the video. I believe this was a big part of the original bug a
couple months ago when I couldn't get WebTransport working and
switched to RSocket and then WebSockets + Muxado. This is an
exciting fix!